requirements.txt
Django==4.2.7
Pillow==10.1.0
django-allauth==0.58.2
django-environ==0.11.2
django-crispy-forms==2.0
crispy-bootstrap5==0.7
sslcommerz-python==0.0.3
django-storages==1.14.2
django-cors-headers==4.3.1
psycopg2-binary==2.9.9
python-decouple==3.8
django-debug-toolbar==4.2.0
django-filter==23.4
django-taggit==4.0.0
Markdown==3.5.1
django-ckeditor==6.7.0
djangorestframework==3.14.0 